# src/karmen_backend/openapi/swagger.yaml
openapi: 3.0.0
info:
  version: '0.9.0'
  title: 'Karmen API'
  description: 'API of https://github.com/fragaria/karmen'
...
paths:
...
  '/organizations/{org_uuid}/gcodes':
    get:
      summary: 'List new gcodes'
      operationId: server.routes.gcodes.gcodes_list
      tags:
        - gcodes
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/limit_count_filter'
        - $ref: '#/components/parameters/order_by_filter'
        - $ref: '#/components/parameters/fields_filter'
        - $ref: '#/components/parameters/filter_filter'
        - in: query
          name: start_with
          schema:
            $ref: '#/components/schemas/uuid_or_positive_integer'
      responses:
        '200':
          description: 'List of gcodes'
    post:
      summary: 'Upload gcode'
      operationId: server.routes.gcodes.gcode_create
      tags:
        - gcodes
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                path:
                  type: string
                  format: binary
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 'Created'
  '/organizations/{org_uuid}/gcodes/{gcode_uuid}':
    get:
      summary: 'Get gcode details'
      operationId: server.routes.gcodes.gcode_detail
      tags:
        - gcodes
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/gcode_uuid_param'
      responses:
        '200':
          description: 'Gcode details'
    delete:
      summary: 'Delete gcode'
      operationId: server.routes.gcodes.gcode_delete
      tags:
        - gcodes
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/gcode_uuid_param'
      responses:
        '204':
          description: 'Deleted'
  '/organizations/{org_uuid}/gcodes/{gcode_uuid}/data':
    get:
      summary: 'Download gcode file'
      operationId: server.routes.gcodes.gcode_file
      tags:
        - gcodes
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/gcode_uuid_param'
      responses:
        '200':
          description: 'Gcode file'
...
  '/organizations/{org_uuid}/printjobs':
    post:
      summary: 'Start printing gcode on a printer'
      operationId: server.routes.printjobs.printjob_create
      tags:
        - printjobs
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                gcode:
                  $ref: '#/components/schemas/uuid'
                printer:
                  $ref: '#/components/schemas/uuid'
      responses:
        '201':
          description: 'Printjob scheduled'
        '409':
          description: Could ot schedule a printjob
    get:
      summary: 'List printjobs'
      operationId: server.routes.printjobs.printjobs_list
      tags:
        - printjobs
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/limit_count_filter'
        - $ref: '#/components/parameters/start_with_filter'
        - $ref: '#/components/parameters/order_by_filter'
        - $ref: '#/components/parameters/fields_filter'
        - $ref: '#/components/parameters/filter_filter'


      responses:
        '200':
          description: 'List of printjobs'
  '/organizations/{org_uuid}/printjobs/{printjob_uuid}':
    get:
      summary: 'Get printjob detail'
      operationId: server.routes.printjobs.printjob_detail
      tags:
        - printjobs
      parameters:
        - $ref: '#/components/parameters/organization_uuid_param'
        - $ref: '#/components/parameters/printjob_uuid_param'
      responses:
        '200':
          description: 'Printjob detail'
...
components:
  parameters:
...
    organization_uuid_param:
      name: org_uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
    printjob_uuid_param:
      name: printjob_uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
...
    gcode_uuid_param:
      name: gcode_uuid
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
...
    limit_count_filter:
      in: query
      name: limit
      schema:
        type: integer
    start_with_filter:
      in: query
      name: start_with
      schema:
        $ref: '#/components/schemas/uuid'
    order_by_filter:
      in: query
      name: order_by
      schema:
        type: string
    fields_filter:
      in: query
      name: fields
      schema:
        type: string
    filter_filter:
      in: query
      name: filter
      schema:
        type: string
  schemas:
    uuid:
      type: string
      format: uuid
      pattern: '[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}'
    uuid_or_positive_integer:
      type: string
      pattern: '^((\d+$)|([0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}))'
...